// src/components/ts/model/member-info-model.ts
export interface DefaultMemberModel {
    /** ID */
    Id: number;
    /** 背番号 */
    Number: number;
    /** 登録名 */
    Name: string;
    /** 行番号 */
    RowNumber: number;
}

/** 
 * 出場選手情報
 */
export class ParticipationMemberModel {
    /** 背番号 */
    Number: number;
    /** 守備位置 */
    Position: string;
    /** 名前 */
    Name: string;
    /** 打者 */
    Batter: boolean;
    /** 走者 */
    Runner: boolean;
    /** 次回先頭 */
    NextRead: boolean;
}

/** 
 * 出場選手情報（1チーム単位）
 */
export class ParticipationMemberPerTeamModel {
    /** 1番 */
    LeadOff: ParticipationMemberModel;
    /** 2番 */
    Second: ParticipationMemberModel;
    /** 3番 */
    Third: ParticipationMemberModel;
    /** 4番 */
    Forth: ParticipationMemberModel;
    /** 5番 */
    Fifth: ParticipationMemberModel;
    /** 6番 */
    Sixth: ParticipationMemberModel;
    /** 7番 */
    Seventh: ParticipationMemberModel;
    /** 8番 */
    Eighth: ParticipationMemberModel;
    /** 9番 */
    Ninth: ParticipationMemberModel;
    /** 投手 */
    Pitcher: ParticipationMemberModel;
}